refactor: KakaoConfig naming 수정
