+++ src/test/java/com/olbimacoojam/heaven/util/KakaoConfigTest.java
@@ -1,8 +1,10 @@
 package com.olbimacoojam.heaven.util;
 
+import com.olbimacoojam.heaven.kakao.KakaoConfig;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.web.reactive.function.client.WebClient;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -13,11 +15,16 @@ class KakaoConfigTest {
 
     @Test
     void KakaoConfig_yml파일_매핑테스트_client() {
-        assertThat(kakaoConfig.getClient().get("clientId")).isNotNull();
+        assertThat(kakaoConfig.getAuth().get("clientId")).isNotNull();
     }
 
     @Test
     void KakaoConfig_yml파일_매핑테스트_resource() {
-        assertThat(kakaoConfig.getResource().get("userInfoUri")).isNotNull();
+        assertThat(kakaoConfig.getResource().get("userInfoPath")).isNotNull();
+    }
+
+    @Test
+    void name() {
+        WebClient webClient = WebClient.create();
     }
 }
